refactor(#343): T130 — extract program draft helpers to _shared/
Move buildProgramPrompt, capCatalog, getEquipmentValues, getExerciseBounds,
validateProgram (+ types) from generate-program/ into _shared/programDraft.ts.
Move callGeminiProgram into _shared/programGemini.ts (split forced by tsc:
Deno.env.get and Array.findLast aren't reachable from vitest specs in src/test/).

generate-program/{prompt,validate,gemini,types}.ts deleted. generate-program/
is now just index.ts — independently deletable by #342's T129 conditional
cleanup with zero cross-epic coordination.

Phase 4 refactor: dedupe CatalogExercise/UserProfile/RecentExercise in
_shared/programCatalog.ts. Its old comment justified the duplication as
"avoiding a feature module #343 will retire" — that justification is now
false since the types live in _shared/. Types re-exported from programDraft.ts;
single source of truth.

Safety net green:
- 244/244 Deno tests pass (supabase/functions/**/*_test.ts)
- 34/34 vitest tests pass (src/test/{validate,prompt}-program.test.ts)
- deno check clean on all three Edge entrypoints
- tsc --noEmit clean

E2E deferred to PR CI per ticket acceptance.
